feat: timestamp diff highlighting (#20)
This marks up the rendered @timestamp field to allow a color scheme to
style the same and different parts independently. The default color scheme
underlines the different part. If the timestamps are in RFD 3339 format it
attempts to detect if the time offset (the time zone) at the end of the
timestamp is the same.

This adds "timestamp", "timestampSame" and "timestampDiff" roles to the
coloring.