feat: add `-i, --include-fields FIELDS` flag for including only certain fields (#15)
This complements `-x FIELDS`. If included field are given, then *only*
those extra fields are shown; with the exception that the title line
is not affected. I.e., `-i FIELDS` do not apply to title line fields. This
feels incongruous with `-x FIELDS`, which *does* apply to
title line fields, but I think it makes for more natural usage.